refactor: move terminal playback scenarios
